requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt --strip-extras requirements-dev.in
#
astroid==3.3.9
    # via pylint
black==25.1.0
    # via -r requirements-dev.in
click==8.1.8
    # via black
colorama==0.4.6
    # via
    #   click
    #   pylint
    #   pytest
coverage==7.8.0
    # via pytest-cov
dill==0.3.9
    # via pylint
flake8==7.2.0
    # via -r requirements-dev.in
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via
    #   -r requirements-dev.in
    #   pylint
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.7
    # via
    #   black
    #   pylint
pluggy==1.5.0
    # via pytest
pycodestyle==2.13.0
    # via flake8
pyflakes==3.3.2
    # via flake8
pylint==3.3.6
    # via -r requirements-dev.in
pytest==8.3.5
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
pytest-cov==6.1.1
    # via -r requirements-dev.in
pytest-django==4.11.1
    # via -r requirements-dev.in
tomlkit==0.13.2
    # via pylint
